fix(worktree): add state_label_dispatch service to code limits and refactor WorktreeManager with PR mixin
