+++ config/settings.yaml
@@ -165,6 +165,9 @@ code_limits:
       - path: "src/vibe3/services/orchestra_status_service.py"
         limit: 450
         reason: "Orchestra 状态查询服务，包含状态快照和格式化逻辑，职责单一但方法较多"
+      - path: "src/vibe3/orchestra/services/state_label_dispatch.py"
+        limit: 450
+        reason: "State label dispatch 服务，包含依赖检查、状态查询、事件发射等紧密耦合的调度逻辑"
 
       # 强耦合测试例外
       - path: "tests/vibe3/agents/test_codeagent_backend.py"
